pyproject.toml
[tool.poetry]
name = "penzoe"
version = "0.1.0"
description = "Engaging and morden way of perserving Tibetan language"
authors = ["tenzin <ten13zin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "8.1.1"
Django = "3.2"
django-heroku = "0.3.1"
gunicorn = "20.0.4"
django-crispy-forms = "1.9.0"
django-allauth = "0.41.0"
PyGithub = "1.51"
django-environ = "0.4.5"
dj-database-url = "^0.5.0"
psycopg2 = "2.8.6"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "2.2.1"
pytest-django = "3.9.0"
django-test-plus = "1.4.0"
flake8 = "3.8.3"
factory-boy = "^3.2.0"
black = "^21.5b1"
isort = "^5.8.0"
pre-commit = "^2.12.1"
coverage = "^5.5"
mypy = "0.782"
pygraphviz = "1.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
